Improve directory mentions
