+++ lib/prompt-editor/src/plugins/atMentions/atMentions.module.css
@@ -1,6 +1,6 @@
 .popover-dimensions {
     margin-top: 20px; /* show on the line below the cursor */
-    width: clamp(200px, 65vw, 440px);
+    width: clamp(300px, 65vw, 440px);
 
     --max-items: 12;
     --mention-item-height: 30px;
