+++ vscode/src/context/openctx/remoteDirectorySearch.ts
@@ -1,7 +1,10 @@
 import { REMOTE_DIRECTORY_PROVIDER_URI } from '@sourcegraph/cody-shared'
 
+import type { Item, Mention } from '@openctx/client'
+import { graphqlClient, isDefined, isError } from '@sourcegraph/cody-shared'
 import { getRepositoryMentions } from './common/get-repository-mentions'
-import { getDirectoryItem, getDirectoryMentions } from './currentRepositoryDirectorySearch'
+import { escapeRegExp } from './remoteFileSearch'
+
 import type { OpenCtxProvider } from './types'
 
 const RemoteDirectoryProvider = createRemoteDirectoryProvider()
@@ -39,6 +42,70 @@ export function createRemoteDirectoryProvider(customTitle?: string): OpenCtxProv
             )
         },
     }
+}
+
+export async function getDirectoryMentions(
+    repoName: string,
+    directoryPath?: string
+): Promise<Mention[]> {
+    const repoRe = `^${escapeRegExp(repoName)}$`
+    const directoryRe = directoryPath ? escapeRegExp(directoryPath) : ''
+    const query = `repo:${repoRe} file:${directoryRe}.*\/.* select:file.directory count:10`
+
+    const dataOrError = await graphqlClient.searchFileMatches(query)
+
+    if (isError(dataOrError) || dataOrError === null) {
+        return []
+    }
+
+    return dataOrError.search.results.results
+        .map(result => {
+            if (result.__typename !== 'FileMatch') {
+                return null
+            }
+
+            const url = `${graphqlClient.endpoint.replace(/\/$/, '')}${result.file.url}`
+
+            return {
+                uri: url,
+                title: result.file.path,
+                description: ' ',
+                data: {
+                    repoName: result.repository.name,
+                    repoID: result.repository.id,
+                    rev: result.file.commit.oid,
+                    directoryPath: result.file.path,
+                },
+            } satisfies Mention
+        })
+        .filter(isDefined)
+}
+
+export async function getDirectoryItem(
+    query: string,
+    repoID: string,
+    directoryPath: string
+): Promise<Item[]> {
+    const dataOrError = await graphqlClient.contextSearch({
+        repoIDs: [repoID],
+        query,
+        filePatterns: [directoryPath],
+    })
+
+    if (isError(dataOrError) || dataOrError === null) {
+        return []
+    }
+
+    return dataOrError.map(
+        node =>
+            ({
+                url: node.uri.toString(),
+                title: node.path,
+                ai: {
+                    content: node.content,
+                },
+            }) as Item
+    )
 }
 
 export default RemoteDirectoryProvider
